+ fix read TranslationName + fix CB_Utils String.Empty and Integer.compare + fix create Store APK + add Settings Value “ShowFieldnotesAsDefaultView”
git-svn-id: svn://svn.code.sf.net/p/droidcachebox/code/branches/2012-11-08%20-%20First%20Release%20on%20Google%20Play@1883 5890f383-b09f-485f-8f59-22bc967d67fd

+++ CB_Utils/src/CB_Utils/Settings/SettingBase.java
@@ -100,7 +100,7 @@ public abstract class SettingBase<T> implements Comparable<SettingBase<T>>
 	@Override
 	public int compareTo(SettingBase<T> o)
 	{
-		return Integer.compare(o.index, this.index);
+		return Double.compare(o.index, this.index);
 	}
 
 	private void fireChangedEvent()
